feat: add a few docker-compose env vars to control the build + docs
Living in docker/compose over the past few days, I figure it'll be good to attach build
args to env vars so it's easy to turn things on and off.

As part of this, I'm turning off installing chromium in the build as in most cases developers
won't be working on something that requires it. If you need though, it's
easy to set an env var to switch it on.

Should speed up the local builds a little and make images significantly
smaller locally.