feat: reduce some filter width in list views
Currently in our CRUD list views (dashboards, charts, datasets) we use a fixed width of 200px for all filters, regardless of their content. This leads to some filter rows overflowing depending on use screen width. Here I'm setting some of the filters to be thinner, preventing filters from wrapping into 2 rows.

In the future version of antd, we'll be able to shrink all filters to thinner width with `popupMatchSelectWidth={false}`, allowing for the dropdown width to be wider than the input box.