fix: GHA frontend builds fail when frontends hasn't changed
I just merged https://github.com/apache/superset/pull/31490 and realize
it fails when no frontend changes are found. It will run the first
workflow and fail on subsequent.

This PR makes it such that the first workflow emits an output to tell
the other workflows to skip if needed.