feat: improve docker-compose services boot sequence
Recently I noticed that despite the service dependencies state in our docker-compose files, the all start on top of each other, which isn't good or as expected. We want for superset-init to finish prior to starting other services, and ideally services being in a healthy state before booting the others. There are newish semantics in docker-compose to enable this, so will implement here in this PR.