feat: cache the fontend's bootstrap data
While reviewing https://github.com/apache/superset/pull/30134 which seemed to make heavy use of `getBootstrapData`, I realized that it's usage is occurs 21 times across the codebase while it may be a fairly expensive operation (JSON.parse on an object that may be significant in size at times (?)).

Now it's easy to cache the result in a closure and ensure it happens only once per SPA-load, but opening this as a DRAFT as we probably already store it in many places.

Tradeoff here is compute VS memory, and clearly we need in memory somewhere, but probably in most cases, already have it in the state app somewhere. Ideally the app would read it once and make it available in its state, as opposed to different modules calling it on-demand...

PR is more of a request for comments.