Fix public paths for assets produced by webpack.
The full public path is only known at runtime. Webpack
supports this through __webpack_public_path__ so it knows
where to load the additional asset chunks.

In addition, the extracted css files have fixed url:src paths to
assets, such as fonts, that they load themselves. The plugin is
now aware of the relative prefix to use for each publicPath
for the extracted css files.