chore: remove all enzyme tests and references
[Enzyme is dead, now
what?](https://dev.to/wojtekmaj/enzyme-is-dead-now-what-ekl)

We're deep in RTL already, and the enzyme tests are doing more harm than good to us as maintainers.

This has implication on lowering test coverage, but at the same time many of the enzyme tests are against fairly settled, untouched components. I suggest that we delete all the Enzyme tests and agree that touching now-untested component requires writing RTL test to cover the components you touch.

Overall a good step forward that allows us to drop a large chunk of complexity and branches in our dep tree.

-----
NOTE: overall there are more enzyme tests than I thought orignally as grepping for `enzyme` missed out on some reusable shims/constructs that indirectly used enzyme. Putting out as DRAFT so we can see what we want to do with this.