fix: set `Rich tooltip` -> 'Show percentage' to false by default
Many users have reported that they prefer having the show percentage option in timeseries as false by default. While it's generally a good feature and many might want it, default should be lightweight and optimize for less clutter.

From my local tests, the false default should affect charts that haven't been saved since the feature was turned on.

Open to input or debatting which default option is best, but given that many users reported it, and that it's a bit of a visual regression / change in some dashboard, I thought we'd switch it off.