Cache entire channel list and do filtering on request
The Slack conversations API is not necessarily faster when using the `types` filter: For example,
when we filter for private channels, the cursor will still iterate over all public channels and the
number of requests remains the same, only the data transmitted is less.

With this in mind, I believe it is more efficient for us to cache only a single value for `get_channels()`
and avoid having multiple cache values for each possible value of `types`. Doing the filtering in
`get_channels_with_search` on demand doesn't add much cost.

Additionally, the `limit` param is also not relevant for the cache key and is inlined.