chore: implement eslint typescript-eslint/consistent-type-imports
Recently looked into typescript's import `type` keyword and its implications. Turns out using it should have noticeable impacts on bundle size and memory usage (yeh!), and there's a nice lint rule for it.

This PR is DRAFT and will be on hold until we merge the Theming branch to avoid a lot of conflicts.

Note that:
- there are north of 2k lint instance
- `--fix` seems to cover most cases, short by 100-200
- unclear how much we'll save on bundle size, but clearly is good hygiene

Let's pick this up after the great merge of https://github.com/apache/superset/pull/31590