fix: `master` builds are failing while trying to push report to cypress
A while back we incurred significant costs while using the "dashboard" reporting feature that cypress exposes, with essentially every CI build costing some money.

At the time we decided to only use the feature on `master`, so that we'd have reports on hand if/when needed and stay under our free quota. Now this part is failing in master, and I don't think anyone is using the feature. This PR disables it for master pushes (same behavior as in PRs). Note that it's still possible (in theory) to manually generate a run and get the report through the `workflow_dispatch` feature, though given the current issue, it would probably require some attention/debugging.

Let's just switch this off to fix master for now. Can always be restored if/when needed.