feat(helm-chart): Allow custom worker image
This change allows setting a custom image for superset worker but
defaulting to the global image configuration set in values.image.

This is useful if you want to run the superset node w/o playwright
and chromium to reduce the image footprint but having a custom
image extension for the worker deployment.