fix: set sideEffect:false in package.json
Was reading recently about how modern tree-shaking 💨🌳🍂 works in modern web apps, and the `sideEffect` property in `package.json` was mentionned so I checked to see our settings.

Apparently we don't have it set right - unless I'm missing something - so I decided to open a DRAFT PR to test things and have the conversation.

Apparently the `sideEffect` we have in `webpack.config.js` is probably an artifact/legacy from a previous era. Nowadays npm packages should define weather/which sideEffect they contain as a hint for the bundler to tree-shake the right things.

Anyhow, let's see if this builds and passes CI.