chore: delete remaining Enzyme tests
Enzyme is long deprecated and we've long moved on to RTL. As discussed with some core committers, we believe it's a net positive to drop all remaining Enzyme tests even if that means lowering test coverage.

Now we benefit from less complexity in the repo, and can focus on test coverage in areas that matter most, without having to deal with the cognitive load of having multiple testing libs, and maintain upgradability to things like React 18 without having to drag Enzyme into that world.